Add Webmachine::Adapter abstract class
Webmachine::Adapter defines the interface for defining a conforming
adapter.
