+++ lib/webmachine.rb
@@ -5,6 +5,7 @@ require 'webmachine/response'
 require 'webmachine/errors'
 require 'webmachine/decision'
 require 'webmachine/streaming'
+require 'webmachine/adapter'
 require 'webmachine/adapters'
 require 'webmachine/dispatcher'
 require 'webmachine/application'
