+++ spec/webmachine/adapters/rack_spec.rb
@@ -39,28 +39,8 @@ describe Webmachine::Adapters::Rack do
     dispatcher.add_route ['test'], Test::Resource
   end
 
-  describe "#initialize" do
-    it "stores the provided configuration" do
-      adapter.configuration.should eql configuration
-    end
-
-    it "stores the provided dispatcher" do
-      adapter.dispatcher.should eql dispatcher
-    end
-  end
-
-  describe ".run" do
-    it "creates a new adapter and runs it" do
-      adapter = mock(described_class)
-
-      described_class.should_receive(:new).
-        with(configuration, dispatcher).
-        and_return(adapter)
-
-      adapter.should_receive(:run)
-
-      described_class.run(configuration, dispatcher)
-    end
+  it "inherits from Webmachine::Adapter" do
+    adapter.should be_a_kind_of(Webmachine::Adapter)
   end
 
   describe "#run" do
